OodleDataCompression fix 64 bit size cast to int
#rb none

[CL 17089616 by charles bloom in ue5-main branch]

+++ Engine/Source/Runtime/Core/Private/Compression/OodleDataCompression.cpp
@@ -128,7 +128,7 @@ struct OodleDataCompressionDecoders
 
 					OodleDecoders[i].OodleDecoderMutex.Unlock();
 
-					return (int) Result;
+					return (int64) Result;
 				}
 
 				OodleDecoders[i].OodleDecoderMutex.Unlock();
